backend/requirements.txt
# Nigerian Payroll Management System - Backend Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.2.3
alembic==1.13.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# CORS and Security
python-multipart==0.0.6

# Date handling
python-dateutil==2.8.2
